On branch develop Ed's review complete. modified:   blog.md
Love the first paragraph!  I like how you set the context and refer the
reader to a more introductory blog post about Open Liberty and security
in general.  This technique absolves you of responsibility to cover the
basics of Open Liberty itself.

> By the way, the sample code used in this blog is hosted on this GitHub
> repo, feel free to check it out and follow its user guide to run the
> demo application before or after reading this blog.

You absolutely can link to the GitHub repo, but we need to capture that
as an activation action.  I'll find out how to do that.

Please see comments in the markdown.

I've made some copyedits as well.